+++ contracts/dapps/payroll/module/examples/schema.rs
@@ -4,8 +4,7 @@ use std::fs::create_dir_all;
 use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
 
 use pandora_os::core::treasury::dapp_base::state::BaseState;
-use payroll_dapp::msg::InstantiateMsg;
-use payroll_dapp::msg::{ExecuteMsg, QueryMsg};
+use pandora_os::dapps::payout::{InstantiateMsg, ExecuteMsg, QueryMsg};
 
 fn main() {
     let mut out_dir = current_dir().unwrap();
