+++ contracts/dapps/payroll/module/src/lib.rs
@@ -2,7 +2,6 @@ mod commands;
 pub mod contract;
 pub mod dapp_base;
 pub mod error;
-pub mod msg;
 pub mod queries;
 pub mod response;
 pub mod state;
