+++ packages/pandora-os/src/dapps/mod.rs
@@ -1,0 +1,1 @@
+pub mod payout;
