+++ packages/pandora-os/src/lib.rs
@@ -5,3 +5,4 @@ pub mod queries;
 pub mod registery;
 pub mod services;
 pub mod util;
+pub mod dapps;
